Fix #51, separate params mosaic_domains_bragg and mosaic_domains_diffuse
Not sure if this is what was requested -- alternative interpretation
would be we need a separate simulator for diffuse scattering entirely
that only generates diffuse signal, and a simulator for Bragg scattering
that only generates Bragg signal. I don't see a good way to internally
keep track of both within a single simulator instance.