Fix #58, only one function call per "big step"
For context: the tk Spinbox widget that we use to control the numerical parameters has the up and down arrow buttons built in, and their behavior cannot be changed. Further, when a particular widget is "active" (the mouse has clicked in the text area, or the user has clicked on the up/down GUI arrows), the up and down keys on the keyboard also automatically trigger the same behavior.

We wanted to produce different behavior when the Shift key was held. Previously I had bound an additional set of events to the shift-up and shift-down keypresses and adjusted the stepsize accordingly so that the total difference was the big stepsize. This involved one unnecessary call to generating image data. To preserve the GUI responding to the GUI Spinbox up and down arrows, we must trigger the increment and decrement function calls of the Spinbox matching the GUI up and down arrow presses. Further, if we bind multiple events to the keyboard up and down arrows (e.g. bind our own function and also let the Spinbox register the change), the Spinbox controls are triggered first, so there is no way to trigger only the Shift-Up keypress without the Up keypress.

It does appear to be possible to modify the increment value itself before the Spinbox executes the increment or decrement. We can set this to the small step size when the Shift key is not held and to the big step size when the Shift key is held. The Shift key being released is not individually tracked as other keyboard events, hence every key release triggering one function call that will continue iff the key released was the Shift key (which counterintuitively is labeled as Caps Lock being released).

Finally, although it would be optimal to only update the increment size for the active control, this (1) may leave controls in an unexpected state if a different widget is selected while the Shift key is held, and (2) will not allow the intended behavior if the Shift key is held before pressing a previously-inactive control's GUI arrow. The current implementation allows the Shift key to behave as expected for any keypress affecting the Spinbox.